+++ features/shadow/distsql/handler/src/main/java/org/apache/shardingsphere/shadow/distsql/handler/query/ShowShadowAlgorithmsExecutor.java
@@ -18,69 +18,56 @@
 package org.apache.shardingsphere.shadow.distsql.handler.query;
 
 import org.apache.shardingsphere.distsql.handler.query.RQLExecutor;
+import org.apache.shardingsphere.infra.config.algorithm.AlgorithmConfiguration;
 import org.apache.shardingsphere.infra.merge.result.impl.local.LocalDataQueryResultRow;
 import org.apache.shardingsphere.infra.metadata.database.ShardingSphereDatabase;
+import org.apache.shardingsphere.infra.util.props.PropertiesConverter;
 import org.apache.shardingsphere.shadow.api.config.ShadowRuleConfiguration;
-import org.apache.shardingsphere.shadow.distsql.parser.statement.ShowShadowTableRulesStatement;
+import org.apache.shardingsphere.shadow.distsql.parser.statement.ShowShadowAlgorithmsStatement;
 import org.apache.shardingsphere.shadow.rule.ShadowRule;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.Iterator;
 import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Optional;
+import java.util.Properties;
 
 /**
- * Show shadow table rule executor.
+ * Show shadow algorithms executor.
  */
-public final class ShowShadowTableRuleExecutor implements RQLExecutor<ShowShadowTableRulesStatement> {
-    
-    private static final String SHADOW_TABLE = "shadow_table";
-    
-    private static final String SHADOW_ALGORITHM_NAME = "shadow_algorithm_name";
+public final class ShowShadowAlgorithmsExecutor implements RQLExecutor<ShowShadowAlgorithmsStatement> {
     
     @Override
-    public Collection<LocalDataQueryResultRow> getRows(final ShardingSphereDatabase database, final ShowShadowTableRulesStatement sqlStatement) {
+    public Collection<LocalDataQueryResultRow> getRows(final ShardingSphereDatabase database, final ShowShadowAlgorithmsStatement sqlStatement) {
         Optional<ShadowRule> rule = database.getRuleMetaData().findSingleRule(ShadowRule.class);
-        Iterator<Map<String, String>> data = Collections.emptyIterator();
-        if (rule.isPresent()) {
-            data = buildData((ShadowRuleConfiguration) rule.get().getConfiguration()).iterator();
+        if (!rule.isPresent()) {
+            return Collections.emptyList();
         }
+        ShadowRuleConfiguration config = (ShadowRuleConfiguration) rule.get().getConfiguration();
+        Iterator<Entry<String, AlgorithmConfiguration>> data = config.getShadowAlgorithms().entrySet().iterator();
+        String defaultAlgorithm = config.getDefaultShadowAlgorithmName();
         Collection<LocalDataQueryResultRow> result = new LinkedList<>();
         while (data.hasNext()) {
-            Map<String, String> row = data.next();
-            result.add(new LocalDataQueryResultRow(row.get(SHADOW_TABLE), row.get(SHADOW_ALGORITHM_NAME)));
+            Entry<String, AlgorithmConfiguration> row = data.next();
+            result.add(new LocalDataQueryResultRow(row.getKey(), row.getValue().getType(), convertToString(row.getValue().getProps()), Boolean.toString(row.getKey().equals(defaultAlgorithm))));
         }
-        return result;
-    }
-    
-    private List<Map<String, String>> buildData(final ShadowRuleConfiguration shadowRuleConfiguration) {
-        List<Map<String, String>> result = new ArrayList<>();
-        shadowRuleConfiguration.getTables().forEach((key, value) -> {
-            Map<String, String> map = new HashMap<>();
-            map.put(SHADOW_TABLE, key);
-            map.put(SHADOW_ALGORITHM_NAME, convertToString(value.getShadowAlgorithmNames()));
-            result.add(map);
-        });
         return result;
     }
     
     @Override
     public Collection<String> getColumnNames() {
-        return Arrays.asList(SHADOW_TABLE, SHADOW_ALGORITHM_NAME);
+        return Arrays.asList("shadow_algorithm_name", "type", "props", "is_default");
     }
     
-    private String convertToString(final Collection<String> shadowTables) {
-        return null == shadowTables ? "" : String.join(",", shadowTables);
+    private String convertToString(final Properties props) {
+        return null != props ? PropertiesConverter.convert(props) : "";
     }
     
     @Override
     public String getType() {
-        return ShowShadowTableRulesStatement.class.getName();
+        return ShowShadowAlgorithmsStatement.class.getName();
     }
 }
