+++ features/shadow/distsql/handler/src/main/resources/META-INF/services/org.apache.shardingsphere.distsql.handler.query.RQLExecutor
@@ -16,7 +16,7 @@
 #
 
 org.apache.shardingsphere.shadow.distsql.handler.query.ShowShadowRuleExecutor
-org.apache.shardingsphere.shadow.distsql.handler.query.ShowShadowTableRuleExecutor
-org.apache.shardingsphere.shadow.distsql.handler.query.ShowShadowAlgorithmExecutor
+org.apache.shardingsphere.shadow.distsql.handler.query.ShowShadowTableRulesExecutor
+org.apache.shardingsphere.shadow.distsql.handler.query.ShowShadowAlgorithmsExecutor
 org.apache.shardingsphere.shadow.distsql.handler.query.ShowDefaultShadowAlgorithmExecutor
 org.apache.shardingsphere.shadow.distsql.handler.query.CountShadowRuleExecutor
